# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
      "AiCredentialReplaceRequest": {
        "additionalProperties": false,
        "description": "The new key. Shape is checked by the gateway, not here.\n\nNo length or pattern constraint is declared on purpose: a failed schema\nconstraint produces a validation error, and validation errors are exactly\nwhere frameworks like to repeat the rejected input.",
        "properties": {
          "api_key": {
            "format": "password",
            "title": "Api Key",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "api_key"
        ],
        "title": "AiCredentialReplaceRequest",
        "type": "object"
      },
      "AiFailureCount": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "title": "Code",
            "type": "string"
          },
          "count": {
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "code",
          "count"
        ],
        "title": "AiFailureCount",
        "type": "object"
      },
      "AiProviderView": {
        "additionalProperties": false,
        "properties": {
          "approved_models": {
            "items": {
              "type": "string"
            },
            "title": "Approved Models",
            "type": "array"
          },
          "budget_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Minor Units"
          },
          "credential_hint": {
            "maxLength": 16,
            "title": "Credential Hint",
            "type": "string"
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "fallback": {
            "items": {
              "type": "string"
            },
            "title": "Fallback",
            "type": "array"
          },
          "last_failure_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Failure At"
          },
          "last_failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Failure Code"
          },
          "last_failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Failure Message"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "models": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Models",
            "type": "object"
          },
          "provider": {
            "pattern": "^[a-z][a-z0-9-]{1,31}$",
            "title": "Provider",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "reserved_minor_units": {
            "minimum": 0.0,
            "title": "Reserved Minor Units",
            "type": "number"
          },
          "spent_minor_units": {
            "minimum": 0.0,
            "title": "Spent Minor Units",
            "type": "number"
          },
          "status": {
            "enum": [
              "HEALTHY",
              "FAILING",
              "UNVERIFIED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "status",
          "credential_hint",
          "region",
          "models",
          "fallback",
          "approved_models",
          "budget_minor_units",
          "currency",
          "spent_minor_units",
          "reserved_minor_units",
          "last_verified_at",
          "last_failure_code",
          "last_failure_message",
          "last_failure_at"
        ],
        "title": "AiProviderView",
        "type": "object"
      },
      "AiSettingsResponse": {
        "additionalProperties": false,
        "properties": {
          "cost_note": {
            "default": "Spend is ResolveFlow's estimate from the client's approved prices, not a balance reported by the provider.",
            "title": "Cost Note",
            "type": "string"
          },
          "fallback_uses": {
            "default": 0,
            "minimum": 0.0,
            "title": "Fallback Uses",
            "type": "integer"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/AiProviderView"
            },
            "title": "Providers",
            "type": "array"
          },
          "recent_failures": {
            "items": {
              "$ref": "#/components/schemas/AiFailureCount"
            },
            "title": "Recent Failures",
            "type": "array"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "providers",
          "recent_failures"
        ],
        "title": "AiSettingsResponse",
        "type": "object"
      },
      "AiVerificationResponse": {
        "additionalProperties": false,
        "properties": {
          "checked_models": {
            "items": {
              "type": "string"
            },
            "title": "Checked Models",
            "type": "array"
          },
          "code": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "provider": {
            "pattern": "^[a-z][a-z0-9-]{1,31}$",
            "title": "Provider",
            "type": "string"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          },
          "stored": {
            "title": "Stored",
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "ok",
          "stored",
          "checked_models"
        ],
        "title": "AiVerificationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefusalResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "schema_version": {
            "const": "v1",
            "default": "v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "RefusalResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Private service contract for grounded support-ticket triage.",
    "title": "ResolveFlow AI Internal API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ai/providers/{provider}/credential": {
      "put": {
        "operationId": "replace_credential_v1_ai_providers__provider__credential_put",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "pattern": "^[a-z][a-z0-9-]{1,31}$",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiCredentialReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiVerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Replace Credential",
        "tags": [
          "ai-settings"
        ]
      }
    },
    "/v1/ai/providers/{provider}/verify": {
      "post": {
        "operationId": "verify_provider_v1_ai_providers__provider__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "pattern": "^[a-z][a-z0-9-]{1,31}$",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiVerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Verify Provider",
        "tags": [
          "ai-settings"
        ]
      }
    },
    "/v1/ai/settings": {
      "get": {
        "operationId": "read_settings_v1_ai_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-membership-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Membership-Id"
            }
          },
          {
            "in": "header",
            "name": "x-resolveflow-role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Resolveflow-Role"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefusalResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Read Settings",
        "tags": [
          "ai-settings"
        ]
      }
    },
...
  }
}